Update project structure and object naming to conform to SwiftLint
